watcher-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
anthropic>=0.40.0
tqdm==4.66.1
python-multipart==0.0.6
pydantic>=2.5.2,<3.0.0
pytest==7.4.3
httpx==0.25.1
PyPDF2==3.0.1
pdfplumber==0.10.3
ftfy==6.1.3
sqlalchemy==2.0.27
alembic==1.13.1
aiosqlite==0.19.0
greenlet==3.0.3
aiofiles==23.2.1
apscheduler==3.10.4
python-socketio==5.11.0
websockets==12.0

# Agentic AI dependencies (pinned for better resolution)
langchain-core==0.3.29
langchain==0.3.13
langchain-google-genai==2.0.8
langgraph==0.2.62  # Optional - can be removed after full migration